+++ app/src/main/java/org/retriever/dailypet/ui/home/care/ModifyCareFragment.kt
@@ -1,4 +1,4 @@
-package org.retriever.dailypet.ui.main
+package org.retriever.dailypet.ui.home.care
 
 import android.annotation.SuppressLint
 import android.os.Bundle
@@ -9,41 +9,39 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.CompoundButton
 import android.widget.Toast
-import androidx.appcompat.widget.AppCompatButton
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.navArgs
 import org.retriever.dailypet.GlobalApplication
 import org.retriever.dailypet.R
-import org.retriever.dailypet.databinding.FragmentAddCareBinding
+import org.retriever.dailypet.databinding.FragmentModifyCareBinding
 import org.retriever.dailypet.model.Resource
-import org.retriever.dailypet.model.main.CareInfo
+import org.retriever.dailypet.model.main.CareModifyInfo
 import org.retriever.dailypet.ui.base.BaseFragment
 import org.retriever.dailypet.util.hideProgressCircular
 import org.retriever.dailypet.util.setViewBackgroundWithoutResettingPadding
 import org.retriever.dailypet.util.showProgressCircular
 
-class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
+class ModifyCareFragment : BaseFragment<FragmentModifyCareBinding>() {
 
     private val homeViewModel by activityViewModels<HomeViewModel>()
     private val jwt = GlobalApplication.prefs.jwt ?: ""
     private var petId = -1
-    private var petName = ""
-    private var dayList: MutableList<Boolean> = mutableListOf(false, false, false, false, false, false, false)
-    private var careList: MutableList<Boolean> = mutableListOf(false, false, false, false, false)
-    private val dayName: List<String> = listOf("SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT")
+    private var careId = -1
     private var careName = ""
+    private var dayList: MutableList<Boolean> = mutableListOf(false, false, false, false, false, false, false)
+    private val dayName: List<String> = listOf("SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT")
     private var SUBMIT = false
 
-    override fun getFragmentBinding(inflater: LayoutInflater, container: ViewGroup?): FragmentAddCareBinding {
-        return FragmentAddCareBinding.inflate(inflater, container, false)
+    override fun getFragmentBinding(inflater: LayoutInflater, container: ViewGroup?): FragmentModifyCareBinding {
+        return FragmentModifyCareBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initProgressCircular()
-        initPostCare()
+        initPatchCare()
         initInfo()
         buttonClick()
         initCheckbox()
@@ -54,8 +52,8 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
         hideProgressCircular(binding.progressCircular)
     }
 
-    private fun initPostCare() = with(binding) {
-        homeViewModel.postCareResponse.observe(viewLifecycleOwner) { event ->
+    private fun initPatchCare() = with(binding) {
+        homeViewModel.patchPetCareResponse.observe(viewLifecycleOwner) { event ->
             event.getContentIfNotHandled()?.let { response ->
                 when (response) {
                     is Resource.Loading -> {
@@ -63,12 +61,12 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
                     }
                     is Resource.Success -> {
                         hideProgressCircular(progressCircular)
-                        Toast.makeText(requireContext(), "케어가 등록되었습니다", Toast.LENGTH_SHORT).show()
+                        Toast.makeText(requireContext(), "케어가 수정되었습니다", Toast.LENGTH_SHORT).show()
                         root.findNavController().popBackStack()
                     }
                     is Resource.Error -> {
                         hideProgressCircular(progressCircular)
-                        Toast.makeText(requireContext(), "케어 등록에 실패하였습니다", Toast.LENGTH_SHORT).show()
+                        Toast.makeText(requireContext(), "케어 수정에 실패하였습니다", Toast.LENGTH_SHORT).show()
                     }
                 }
             }
@@ -77,38 +75,18 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
 
     @SuppressLint("SetTextI18n")
     private fun initInfo() {
-        val args: AddCareFragmentArgs by navArgs()
+        val args: ModifyCareFragmentArgs by navArgs()
         petId = args.petId
-        petName = args.petName
+        careId = args.careId
+        careName = args.careName
         val title = binding.addCareTitleText.text
-        binding.addCareTitleText.text = "$petName $title"
+        binding.addCareTitleText.text = "$careName $title"
     }
 
     private fun buttonClick() = with(binding) {
-        btnFood.setOnClickListener {
-            selectCare(btnFood, 1)
-            submitCheck()
-        }
-        btnWalk.setOnClickListener {
-            selectCare(btnWalk, 2)
-            submitCheck()
-        }
-        btnPlay.setOnClickListener {
-            selectCare(btnPlay, 3)
-            submitCheck()
-        }
-        btnWash.setOnClickListener {
-            selectCare(btnWash, 4)
-            submitCheck()
-        }
-        btnInputCare.setOnClickListener {
-            careInputEdittext.visibility = View.VISIBLE
-            selectCare(btnInputCare, 0)
-
-        }
         addCareSubmitButton.setOnClickListener {
             if (SUBMIT) {
-                postCare()
+                patchCare()
             }
         }
         backButton.setOnClickListener {
@@ -116,31 +94,20 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
         }
     }
 
-    private fun postCare() {
+    private fun patchCare() {
         val list: MutableList<String> = mutableListOf()
         for (i in 0 until 7) {
             if (dayList[i]) {
                 list.add(dayName[i])
             }
         }
-        val careInfo = CareInfo(
-            careName = careName,
+        val careModifyInfo = CareModifyInfo(
             dayOfWeeks = list,
             totalCountPerDay = binding.careCountEdittext.text.toString().toInt()
         )
-        homeViewModel.postCare(petId, jwt, careInfo)
+        homeViewModel.patchPetCare(petId, careId, jwt, careModifyInfo)
     }
 
-    private fun selectCare(selectedBtn: AppCompatButton, idx : Int) = with(binding) {
-        disableButton()
-        if(selectedBtn != btnInputCare){
-            careInputEdittext.visibility = View.GONE
-        }
-        careList[idx] = true
-        selectedBtn.isSelected = true
-        careName = selectedBtn.text.toString()
-        submitCheck()
-    }
 
     private fun initCheckbox() = with(binding) {
         val listener = CompoundButton.OnCheckedChangeListener { buttonView, isChecked ->
@@ -179,17 +146,6 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
         checkSun.setOnCheckedChangeListener(listener)
     }
 
-    private fun disableButton() = with(binding){
-        for(i in 0 until careList.size){
-            careList[i] = false
-        }
-        btnFood.isSelected = false
-        btnWalk.isSelected = false
-        btnPlay.isSelected = false
-        btnWash.isSelected = false
-        btnInputCare.isSelected = false
-    }
-
     private fun allCheck(check: Boolean) = with(binding){
         for(i in 0 until dayList.size){
             dayList[i] = false
@@ -216,30 +172,13 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
             override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
             }
         })
-        careInputEdittext.addTextChangedListener(object : TextWatcher {
-            override fun afterTextChanged(p0: Editable?) {
-                submitCheck()
-                if (careInputEdittext.text.isNotBlank()) {
-                    careInputEdittext.setViewBackgroundWithoutResettingPadding(R.drawable.whiteblue_click_button)
-                }
-            }
-            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
-            }
-            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
-            }
-        })
     }
 
     private fun submitCheck() = with(binding){
         val day = dayList.contains(true)
-        var care = careList.contains(true)
         val number = careCountEdittext.text.isNotBlank()
-        if(careList[0]){
-            care = careInputEdittext.text.isNotBlank()
-            careName = careInputEdittext.text.toString()
-        }
 
-        SUBMIT = day && care && number
+        SUBMIT = day && number
         if (SUBMIT) {
             addCareSubmitButton.background = ContextCompat.getDrawable(requireContext(), R.drawable.blue_button)
             addCareSubmitButton.setTextColor(ContextCompat.getColor(requireContext(), R.color.white))
