+++ app/src/main/java/org/retriever/dailypet/ui/home/statistics/adapter/GroupAdapter.kt
@@ -1,4 +1,4 @@
-package org.retriever.dailypet.ui.statistics.adapter
+package org.retriever.dailypet.ui.home.statistics.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
