+++ app/src/main/res/layout/fragment_diary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.HomeFragment">
+    tools:context=".ui.home.HomeFragment">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
