+++ app/src/main/res/navigation/main_navigation_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="org.retriever.dailypet.ui.main.HomeFragment"
+        android:name="org.retriever.dailypet.ui.home.HomeFragment"
         android:label="HomeFragment"
         tools:layout="@layout/fragment_home" />
